Code overhaul for improved data handling
There has been a thorough change in the structure of the code. I have included code to generate a pickle file for the csv that directly loads the numpy matrix from disk without having to read a CSV everytime, thus making things a LOT faster.

A consequence of these changes is that once everyone has generated their pickle files they can completely discard the csv.

Moving forward, the three teams do not have to bother with handling the data directly anymore but will only call a handful of functions to avail all the data they will need to design and train their neural networks.

Jitihn and Karan's teams only need to call the function getChangeVolData in PeriodSample Class inside Formatter file. This will give you the price change and trading volume data for the past N periods and the change over the next 1 period in the form of which bin it lies in numbered 0-7 as described in the project proposal

Please let me know before making any changes in the core code files